apps/worker/pyproject.toml
[tool.poetry]
name = "infra-insight-worker"
version = "0.1.0"
description = "Infra Insight MVP worker"
authors = ["You <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
celery = "^5.4.0"
redis = "^5.0.6"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.4.0"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
kubernetes = "^30.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
